Move utility classes into util package
